Add bool for checking the connection state

+++ AirCombatMatchmakerBot/BotManagement/BotRuntime.cs
@@ -22,49 +22,56 @@ public class BotRuntimeManager
 
         client.Ready += async () =>
         {
-            BotReference.SetConnectionState(true);
-            Log.WriteLine("Bot is connected!", LogLevel.DEBUG);
-
-            //ONLY FOR TESTING, DELETES ALL CHANNELS AND CATEGORIES
-            
-            /*
-            var guild = BotReference.GetGuildRef();
-            foreach (var ch in guild.Channels)
+            if (!BotReference.GetConnectionState())
             {
-                if (ch.Name == "info") continue;
+                BotReference.SetConnectionState(true);
+                Log.WriteLine("Bot is connected!", LogLevel.DEBUG);
 
-                Log.WriteLine("deleting " + ch.Name, LogLevel.DEBUG);
-                await ch.DeleteAsync();
+                //ONLY FOR TESTING, DELETES ALL CHANNELS AND CATEGORIES
+
+                /*
+                var guild = BotReference.GetGuildRef();
+                foreach (var ch in guild.Channels)
+                {
+                    if (ch.Name == "info") continue;
+
+                    Log.WriteLine("deleting " + ch.Name, LogLevel.DEBUG);
+                    await ch.DeleteAsync();
+                }
+                foreach (var cat in guild.CategoryChannels)
+                {
+                    Log.WriteLine("deleting category: " + cat.Name, LogLevel.DEBUG);
+                    await cat.DeleteAsync();
+                }*/
+
+                // Creates the league references to the database
+                // (must be run before creating the channels)
+                await LeagueManager.CreateLeaguesOnStartupIfNecessary();
+
+                // Creates the categories and the channels from the interfaces
+                // (dependant on the data from CreateLeaguesOnStartupIfNecessary())
+                await CategoryAndChannelManager.CreateCategoriesAndChannelsForTheDiscordServer();
+
+                // Checks the users that left during down time and sets their teams active
+                await DowntimeManager.CheckForUsersThatLeftDuringDowntime();
+                await SerializationManager.SerializeUsersOnTheServer();
+
+                client.UserJoined += UserManager.HandleUserJoin;
+                client.ButtonExecuted += ButtonHandler.HandleButtonPress;
+                //client.MessageUpdated += UserManager.MessageUpdated;
+
+                // If a member's nickname changes
+                client.GuildMemberUpdated += Database.Instance.PlayerData.HandleRegisteredMemberUpdated;
+                client.UserLeft += UserManager.HandleUserLeaveDelegate;
+
+                //client.ChannelCreated += ChannelManager.FinishChannelCreationFromDelegate;
+
+                await DowntimeManager.CheckForUsersThatJoinedAfterDowntime();
             }
-            foreach (var cat in guild.CategoryChannels)
+            else
             {
-                Log.WriteLine("deleting category: " + cat.Name, LogLevel.DEBUG);
-                await cat.DeleteAsync();
-            }*/
-
-            // Creates the league references to the database
-            // (must be run before creating the channels)
-            await LeagueManager.CreateLeaguesOnStartupIfNecessary();
-
-            // Creates the categories and the channels from the interfaces
-            // (dependant on the data from CreateLeaguesOnStartupIfNecessary())
-            await CategoryAndChannelManager.CreateCategoriesAndChannelsForTheDiscordServer();
-
-            // Checks the users that left during down time and sets their teams active
-            await DowntimeManager.CheckForUsersThatLeftDuringDowntime();
-            await SerializationManager.SerializeUsersOnTheServer();
-
-            client.UserJoined += UserManager.HandleUserJoin;
-            client.ButtonExecuted += ButtonHandler.HandleButtonPress;
-            //client.MessageUpdated += UserManager.MessageUpdated;
-
-            // If a member's nickname changes
-            client.GuildMemberUpdated += Database.Instance.PlayerData.HandleRegisteredMemberUpdated;
-            client.UserLeft += UserManager.HandleUserLeaveDelegate;
-
-            //client.ChannelCreated += ChannelManager.FinishChannelCreationFromDelegate;
-
-            await DowntimeManager.CheckForUsersThatJoinedAfterDowntime();
+                Log.WriteLine("Bot was already connected!", LogLevel.WARNING);
+            }
         };
 
         // Listens for the commandService
